fix compiler warnings
git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@1942 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/libopensc/pkcs15-sec.c
@@ -138,7 +138,7 @@ int sc_pkcs15_decipher(struct sc_pkcs15_card *p15card,
 
 	/* Strip any padding */
 	if (pad_flags & SC_ALGORITHM_RSA_PAD_PKCS1) {
-		r = sc_pkcs1_strip_02_padding(out, r, out, (size_t *) &r);
+		r = sc_pkcs1_strip_02_padding(out, (size_t)r, out, (size_t *) &r);
                 SC_TEST_RET(ctx, r, "Invalid PKCS#1 padding");
 	}
